rename "xtask build" to "xtask self-build"
