+++ xtask/src/self_build.rs
